fix scripts

+++ download.sh
@@ -1,12 +1,6 @@
 model_dir="/workspace/models/Stable-diffusion"
 
-if [ $(ls -A "${model_dir}" | wc -l) -gt 1 ];
-then
-  echo "Model Downloader: Model dir is not empty, aborted..."
-  exit
-fi
-
-echo "Model Downloader: Download started."
+echo "[Model Downloader] Download started."
 
 declare -A models=(
   ["waiNSFWIllustrious_v120.safetensors"]="https://civitai.com/api/download/models/1490781"
@@ -14,9 +8,16 @@ declare -A models=(
 
 for model in ${!models[@]};
 do
-  wget ${models[${model}]} -qO "${model_dir}/${model}.tmp"
+  if [[ -e "${model_dir}/${model}" ]]; then
+    echo "[Model Downloader] Skip: ${model}"
+    continue
+  fi
+
+  echo "[Model Downloader] Downloading: ${model}"
+  wget ${models[${model}]} -qcO "${model_dir}/${model}.tmp"
   sync
   mv "${model_dir}/${model}.tmp" "${model_dir}/${model}"
+  echo "[Model Downloader] Downloaded: ${model}"
 done
 
-echo "Model Downloader: Download completed."
+echo "[Model Downloader] Download completed."
